# Conflicts:
#	app/build.gradle
#	app/src/main/java/shipdoandem/amytateam/org/shipdoandem/activities/MainActivity.java
#	app/src/main/java/shipdoandem/amytateam/org/shipdoandem/fragment/MostViewFragment.java
#	app/src/main/res/layout/fragment_most_view.xml
#	app/src/main/res/layout/item_food.xml

+++ app/src/main/java/shipdoandem/amytateam/org/shipdoandem/databases/models/FoodRespon.java
@@ -13,7 +13,7 @@ public class FoodRespon {
     @SerializedName("cout_rate")
     private int coutRate;
     @SerializedName("rate")
-    private int rate;
+    private float rate;
     @SerializedName("coint_new")
     private String cointNew;
     @SerializedName("_id")
@@ -52,7 +52,7 @@ public class FoodRespon {
         this.coutRate = coutRate;
     }
 
-    public int getRate() {
+    public float getRate() {
         return rate;
     }
